Turned Window class into Singleton - now using Window abilities like drawing etc. is much more comfortable, there's no need to send Window instance to the states' constructors and it's avaliable everywhere.

+++ Window.cpp
@@ -20,6 +20,6 @@ void Window::startRender() { window.clear(); }
 
 void Window::finishRender() { window.display(); }
 
-int Window::getX() { return x; }
+float Window::getX() { return x; }
 
-int Window::getY() { return y; }
+float Window::getY() { return y; }
